+++ src/Internal/Minor.js
@@ -1,5 +1,4 @@
 const DOMCache = new Map()
-
 const DOM = (id) => {
     const cachedEl = DOMCache.get(id)
     if (cachedEl) return cachedEl
@@ -7,4 +6,8 @@ const DOM = (id) => {
     const el = document.getElementById(id)
     DOMCache.set(id, el)
     return el
+}
+
+function getRandom(min, max) {
+    return Math.floor(Math.random() * (max - min) ) + min
 }