+++ src/Main.js
@@ -3,8 +3,8 @@ function increaseOddities(i){
 }
 function calculateOddityGain(){
     data.oddityGain =
-        data.derivs[0].amt.times(theoryEffects[0]).times(upgradeEffects[0]).times(theoryEffects[2]).times(theoryEffects[7]).times(theoryEffects[9])
-            .times(dreamParticleEffects[0]).times(theoryEffects[10]).times(dreamParticleEffects[2])
+        (data.derivs[0].amt.times(CYCLES[1].effect())).times(theoryEffects[0]).times(upgradeEffects[0]).times(theoryEffects[2]).times(theoryEffects[7])
+            .times(theoryEffects[9]).times(dreamParticleEffects[0]).times(theoryEffects[10]).times(dreamParticleEffects[2]).times(theoryEffects[18])
 }
 let diff
 function mainLoop(){
@@ -15,6 +15,7 @@ function mainLoop(){
     produceDerivs(diff)
     increaseOddities(data.oddityGain.times(diff))
     automate()
+    if(data.hasLostTheory[3]) for(let i=0;i<data.circleProg.length;i++) progress(i, data.circleProg[i])
     updateHTML()
 }
 function calculationsLoop(){