+++ src/SmallScripts/MyMath.js
@@ -1,3 +1,0 @@
-function getRandom(min, max) {
-    return Math.floor(Math.random() * (max - min) ) + min
-}