+++ changelog.md
@@ -2,7 +2,7 @@
 (because I love changelogs) <br>
 Only changes I consider notable are recorded here, no bugfixes or small changes.
 ### Most Recent
-**v0.0.13** [The Great Rollback]: Removed the Shop. Code Optimizations. <br>
+**v0.0.19** [The Circular Update]: The Circles, 9 Cycles, 4 Breakpoints, 4 Secrets, 5 new Theories, 1 new Lost Theory, and 2 new Lost Cycles! <br>
 ### Pre-Alpha
 ***
 **v0.0.1** Created this repo, created the base files, added Saving, Loading, Import, and Export systems, and added a favicon. <br>
@@ -28,3 +28,5 @@ Only changes I consider notable are recorded here, no bugfixes or small changes.
 **v0.0.15b** More content is hidden until you get close to it. <br>
 **v0.0.16** [The Tutorial with Infinite Steps]: The Infinite Stairway, 8 Stair Secrets, Lost Cycle 3, and Legend 3! <br>
 **v0.0.17** [The Singular Update]: The Blackhole, Stair Secret 8, 6 Ringularities, and 2 new Theories! <br>
+**v0.0.18** [The SECOND Rollback]: Reverted to v0.0.15b <br>
+**v0.0.19** [The Circular Update]: The Circles, 9 Cycles, 4 Breakpoints, 4 Secrets, 5 new Theories, 1 new Lost Theory, and 2 new Lost Cycles! <br>